* 'master' of ssh://digitalcave.ca/~/projects:
  turn the motors ALL on!
  Drum enhancements:
  Temporarily remove fadeout, as it was causing problems.  Removed support for random samples as random() doesnt seem to work properly (?!??!?).  Things are looking pretty good now... uploaded a sample video from this current state.
  Misc fixes
  Fade audio out when muting cymbals.
  Updated schematic + pin definitions to have the motor numbers in the proper order, and added a document indicating configuration.
  deleted old frame diagram
  Improve startup by showing Loading mask, and initializing the display prior to TWI / SPI
  Fixed some bugs in the new double-math volume department...
  Pass doubles around for volume instead of ints, to avoid overflow and rounding errors
  Drain
  Multiple changes which will hopefully increase stability: -Move display initialization to be explicit, after proper startup routines (should prevent the display from erroring out on startup) -Use 5x Mixer4 rather than one Mixer16.  Paul indicates that using large mixers causes excessive memory usage.  Not sure if I will keep this change, I may revert it...
  Updated schematic / board with proper voltage divider values
  Fixed some integer overflows, and adjusted timings to increase read stability.
  Updated Teensy libraries to latest versions.  Had to do a few modifications (specifically, add Mixer16 and keep support for configurable CS pin on SPI Flash chip.  Code compiles but is untested.
  Reduced sample count from 12 to 10 (reduce system load), and add 10 microsecond delays after switching mux channels / enabling a mux.  This seems to give better results.
  Prevent overflow on adc value
  Improved signal stability... I think...

+++ projects/drummaster/rev2/src/pad/HiHat.cpp
@@ -11,7 +11,7 @@ char* HiHat::lookupFilename(uint8_t volume){
 }
 
 void HiHat::poll(){
-	uint8_t volume = readPiezo(piezoMuxIndex);
+	double volume = readPiezo(piezoMuxIndex);
 	if (volume){
 		play(volume);
 	}
